% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author.R
\name{author}
\alias{author}
\title{Query author from the Goodreads API}
\usage{
author(id)
}
\arguments{
\item{id}{Goodreads ID of the author}
}
\value{
A \code{data.frame} of author with corresponding information
}
\description{
Query for Goodreads authors based on IDs.
}

