% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_by_name.R
\name{author_by_name}
\alias{author_by_name}
\title{Query author by name from the Goodreads API}
\usage{
author_by_name(name)
}
\arguments{
\item{name}{Name of the author}
}
\value{
A \code{data.frame} of author with corresponding information
}
\description{
Query for Goodreads authors based on names.
}

