% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{book}
\alias{book}
\title{Query book from the Goodreads API}
\usage{
book(id)
}
\arguments{
\item{id}{Goodreads ID of the book}
}
\value{
A \code{data.frame} of book with corresponding information
}
\description{
Query for Goodreads books based on IDs.
}

