% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_by_isbn.R
\name{book_by_isbn}
\alias{book_by_isbn}
\title{Query book from the Goodreads API by ISBN}
\usage{
book_by_isbn(isbn)
}
\arguments{
\item{isbn}{ISBN of the book}
}
\value{
A \code{data.frame} of book with corresponding information
}
\description{
Query for Goodreads books based on ISBN.
}

