% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_by_title.R
\name{book_by_title}
\alias{book_by_title}
\title{Query book from the Goodreads API by title}
\usage{
book_by_title(title)
}
\arguments{
\item{title}{Title of the book}
}
\value{
A \code{data.frame} of book with corresponding information
}
\description{
Query for Goodreads books based on titles.
}

