% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{goodreads_GET}
\alias{goodreads_GET}
\title{Parse the results of a Goofreads API query into a data.frame.}
\usage{
goodreads_GET(path, site = "goodreads", page = 1, num_pages = 1, ...)
}
\arguments{
\item{req}{a request from httr::GET}
}
\description{
Parse the results of a Goofreads API query into a data.frame.
}

