% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_search.R
\name{group_search}
\alias{group_search}
\title{Search for groups from the Goodreads API}
\usage{
group_search(q)
}
\arguments{
\item{q}{Query term for group search}
}
\value{
A \code{data.frame} of groups with corresponding information
}
\description{
Search for groups based on query terms.
}

