% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_reviews.R
\name{recent_reviews}
\alias{recent_reviews}
\title{Query recent reviews from the Goodreads API}
\usage{
recent_reviews()
}
\value{
A \code{data.frame} of reviews with corresponding information
}
\description{
Query for Goodreads recent book reviews.
}

