% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\title{Query review from the Goodreads API}
\usage{
review(id)
}
\arguments{
\item{id}{Goodreads ID of the review}
}
\value{
A \code{data.frame} of review with corresponding information
}
\description{
Query for Goodreads reviews based on IDs.
}

