% Generated by hinterland-1: do not edit by hand
% Please edit documentation in R/recent_reviews.R
\name{review_by_book}
\alias{review_by_book}
\title{Query reviews by book id from the Goodreads API}
\usage{
review_by_book(id,rating,text_only)
}
\arguments{
\item{id}{Goodreads ID of the Book}
}
\value{
A \code{data.frame} of reviews with corresponding information
}
\description{
Query for Goodreads book reviews from book id.
}
