% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{Query users from the Goodreads API}
\usage{
user(id)
}
\arguments{
\item{id}{User ID.}
}
\value{
A \code{data.frame} of the user with corresponding information
}
\description{
Query for Goodreads user by IDs.
}

